+++ packages/gestalt/src/Masonry/__integration__/resizeReflow.integration.js
@@ -1,7 +1,5 @@
-/* global describe */
-/* global it */
 import assert from 'assert';
-import ghost from 'ghostjs';
+
 import countColumns from './lib/countColumns';
 import triggerResize from './lib/triggerResize';
 
@@ -13,31 +11,30 @@ describe('Masonry > Resize', () => {
 
     // This test cares about page size, so close the previous instance to ensure
     // we open a new window with the correct dimensions.
-    ghost.close();
-    await ghost.open('http://localhost:3001/Masonry', {
-      viewportSize: {
-        width: GRID_WIDTH,
-        height: 800,
-      },
+
+    await page.setViewport({
+      width: GRID_WIDTH,
+      height: 800,
     });
+    await page.goto('http://localhost:3001/Masonry');
 
     // Wait for the grid to be hydrated.
     // TODO: Break this out into a utility /w wait() instead.
-    await ghost.wait(1000);
+    await page.waitFor(1000);
 
     const expectedColumns = Math.floor(GRID_WIDTH / PIN_SIZE);
     assert.equal(
-      await countColumns(),
+      await countColumns(page),
       expectedColumns,
       `expected ${expectedColumns} columns`
     );
 
-    await triggerResize(GRID_WIDTH - PIN_SIZE);
+    await triggerResize(GRID_WIDTH - PIN_SIZE, page);
 
     // Wait for the resize measurements to be finished
-    await ghost.wait(() => ghost.script(() => window.RESIZE_MEASUREMENT_DONE));
+    await page.waitForFunction(() => window.RESIZE_MEASUREMENT_DONE);
     assert.equal(
-      await countColumns(),
+      await countColumns(page),
       expectedColumns - 1,
       `expected ${expectedColumns - 1} columns after resize`
     );