+++ packages/gestalt/src/Masonry/__integration__/shuffleItems.integration.js
@@ -1,23 +1,17 @@
-/* global describe */
-/* global it */
 import assert from 'assert';
-import ghost from 'ghostjs';
 import selectors from './lib/selectors';
 
 const getItemColumnMap = async () => {
-  const gridItems = await ghost.findElements(selectors.gridItem);
+  const gridItems = await page.$$(selectors.gridItem);
   const itemLeftMap = {};
   for (let i = 0; i < gridItems.length; i += 1) {
-    const isVisible = await gridItems[i].isVisible();
-    if (isVisible) {
-      const itemRect = await gridItems[i].rect();
-      itemLeftMap[itemRect.left] = itemLeftMap[itemRect.left] || [];
-      itemLeftMap[itemRect.left].push({
-        ...itemRect,
-        itemIndex: i,
-        text: await gridItems[i].text(),
-      });
-    }
+    const boundingBox = await gridItems[i].boundingBox();
+    itemLeftMap[boundingBox.x] = itemLeftMap[boundingBox.x] || [];
+    itemLeftMap[boundingBox.x].push({
+      ...boundingBox,
+      itemIndex: i,
+      text: (await gridItems[i].getProperty('innerText')).toString(),
+    });
   }
 
   return itemLeftMap;
@@ -25,21 +19,19 @@ const getItemColumnMap = async () => {
 
 describe('Masonry > Shuffle items', () => {
   it('Should reflow the grid when items are shuffled ', async () => {
-    ghost.close();
-    await ghost.open('http://localhost:3001/Masonry', {
-      viewportSize: {
-        width: 800,
-        height: 800,
-      },
+    await page.setViewport({
+      width: 800,
+      height: 800,
     });
+    await page.goto('http://localhost:3001/Masonry');
 
     const originalItemMap = await getItemColumnMap();
 
-    const insertTrigger = await ghost.findElement(selectors.shufflePins);
+    const insertTrigger = await page.$(selectors.shufflePins);
     await insertTrigger.click();
 
+    await page.waitFor(50);
     const newItemMap = await getItemColumnMap();
-
     // assert that the first row of items has changed
     // todo (yen) - this can be a more thorough test
     Object.keys(originalItemMap).forEach(col => {