+++ test/containers/MasonryExample.js
@@ -76,7 +76,6 @@ export default class MasonryExample extends React.Component {
     // set flag once masonry mounts - used for remount regression testing
     window.MASONRY_DID_MOUNT = true;
   }
-
   getItems = ({ name = 'Item', from = 0 }, collage) => {
     let until = from + 20;
 
@@ -98,7 +97,6 @@ export default class MasonryExample extends React.Component {
       }
       window.TEST_FETCH_COUNTS = window.TEST_FETCH_COUNTS || 0;
       window.TEST_FETCH_COUNTS += 1;
-
       window.NEXT_FETCH = () => {
         window.NEXT_FETCH = null;
         resolve(items);